Seperated the domain ability from the rat king component.
