+++ Content.Client/RatKing/RatKingDomainSystem.cs
@@ -1,0 +1,9 @@
+using Content.Shared.RatKing;
+
+namespace Content.Client.RatKing;
+
+/// <inheritdoc/>
+public sealed class RatKingDomainSystem : SharedRatKingDomainSystem
+{
+
+}
