+++ Content.Shared/RatKing/RatKingComponent.cs
@@ -5,27 +5,8 @@ using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototy
 
 namespace Content.Shared.RatKing;
 
-[RegisterComponent, NetworkedComponent, Access(typeof(SharedRatKingSystem))]
+[RegisterComponent, NetworkedComponent]
 public sealed partial class RatKingComponent : Component
 {
-    [DataField("actionDomain", customTypeSerializer: typeof(PrototypeIdSerializer<EntityPrototype>))]
-    public string ActionDomain = "ActionRatKingDomain";
 
-    /// <summary>
-    ///     The action for the Domain ability
-    /// </summary>
-    [DataField("actionDomainEntity")]
-    public EntityUid? ActionDomainEntity;
-
-    /// <summary>
-    ///     The amount of hunger one use of Domain consumes
-    /// </summary>
-    [DataField("hungerPerDomainUse", required: true), ViewVariables(VVAccess.ReadWrite)]
-    public float HungerPerDomainUse = 50f;
-
-    /// <summary>
-    ///     How many moles of ammonia are released after one us of Domain
-    /// </summary>
-    [DataField("molesAmmoniaPerDomain"), ViewVariables(VVAccess.ReadWrite)]
-    public float MolesAmmoniaPerDomain = 200f;
 }
