+++ Content.Shared/RatKing/RatKingDomainComponent.cs
@@ -1,0 +1,31 @@
+using Robust.Shared.GameStates;
+using Robust.Shared.Prototypes;
+using Robust.Shared.Serialization;
+using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
+
+namespace Content.Shared.RatKing;
+
+[RegisterComponent, NetworkedComponent, Access(typeof(SharedRatKingDomainSystem))]
+public sealed partial class RatKingDomainComponent : Component
+{
+    [DataField("actionDomain", customTypeSerializer: typeof(PrototypeIdSerializer<EntityPrototype>))]
+    public string ActionDomain = "ActionRatKingDomain";
+
+    /// <summary>
+    ///     The action for the Domain ability
+    /// </summary>
+    [DataField("actionDomainEntity")]
+    public EntityUid? ActionDomainEntity;
+
+    /// <summary>
+    ///     The amount of hunger one use of Domain consumes
+    /// </summary>
+    [DataField("hungerPerDomainUse", required: true), ViewVariables(VVAccess.ReadWrite)]
+    public float HungerPerDomainUse = 50f;
+
+    /// <summary>
+    ///     How many moles of ammonia are released after one us of Domain
+    /// </summary>
+    [DataField("molesAmmoniaPerDomain"), ViewVariables(VVAccess.ReadWrite)]
+    public float MolesAmmoniaPerDomain = 200f;
+}
