+++ Content.Shared/RatKing/SharedRatKingDomainSystem.cs
@@ -1,0 +1,46 @@
+using Content.Shared.Actions;
+using Content.Shared.DoAfter;
+using Content.Shared.Random;
+using Content.Shared.Random.Helpers;
+using Content.Shared.Verbs;
+using Robust.Shared.Audio;
+using Robust.Shared.Audio.Systems;
+using Robust.Shared.Network;
+using Robust.Shared.Prototypes;
+using Robust.Shared.Random;
+using Robust.Shared.Serialization;
+
+namespace Content.Shared.RatKing;
+
+public abstract class SharedRatKingDomainSystem : EntitySystem
+{
+    [Dependency] private readonly INetManager _net = default!;
+    [Dependency] protected readonly IPrototypeManager PrototypeManager = default!;
+    [Dependency] protected readonly IRobustRandom Random = default!;
+    [Dependency] private readonly SharedActionsSystem _action = default!;
+    [Dependency] private readonly SharedAudioSystem _audio = default!;
+    [Dependency] private readonly SharedDoAfterSystem _doAfter = default!;
+
+    /// <inheritdoc/>
+    public override void Initialize()
+    {
+        SubscribeLocalEvent<RatKingDomainComponent, ComponentStartup>(OnStartup);
+        SubscribeLocalEvent<RatKingDomainComponent, ComponentShutdown>(OnShutdown);
+    }
+
+    private void OnStartup(EntityUid uid, RatKingDomainComponent component, ComponentStartup args)
+    {
+        if (!TryComp(uid, out ActionsComponent? comp))
+            return;
+
+        _action.AddAction(uid, ref component.ActionDomainEntity, component.ActionDomain, component: comp);
+    }
+
+    private void OnShutdown(EntityUid uid, RatKingDomainComponent component, ComponentShutdown args)
+    {
+        if (!TryComp(uid, out ActionsComponent? comp))
+            return;
+
+        _action.RemoveAction(uid, component.ActionDomainEntity, comp);
+    }
+}
